auth global options
